feat: integrate parallel thinking optimizer
