+++ tests/test_cache.py
@@ -1,4 +1,3 @@
-import os
 import sys
 import types
 import asyncio
